Docker development file

+++ setup/rebuild.sh
@@ -1,4 +1,13 @@
 #!/bin/bash
-cd ..
+
+if [[ "${PWD##*/}" != "fondant-app" ]]; then
+    echo "Please execute this script in the fondant-app folder."
+    exit 1
+fi
+
+
 docker-compose down
-docker-compose up --build
+
+docker-compose build
+
+docker-compose up
